requirements.txt
colorlog==6.8.2
homeassistant==2024.6.0
pip>=21.3.1
ruff==0.7.0
py-my-bezeq==0.1.3
